Renamed some 2 model files and removed duplicate model file
